Random Forest Code Complete
As we searching the Random Forest methods, we found a paper under same team. The paper name is [Cross-Project Transfer Representation Learning for Vulnerable Function Discovery (https://ieeexplore.ieee.org/abstract/document/8329207/).
They are using same functions to identify the vulnerable and unvulnerable functions which is examine the keyword "cve" and "CVE".

The code is executable with POSTER dataset

Co-Authored-By: zhangfanfansz <58411513+zhangfanfansz@users.noreply.github.com>