Fixed Random Forest Method
After reviewing our process and compare with the paper process, we found that the utilization of source code data does not match the paper description. The paper described a proxy tool (Understand) for extracting the code  metric. However, our approach used the extracted AST from codesenor as representations
and trained with Random Forest Algorithm which explains the differences between our  two attempts.

Co-Authored-By: zhangfanfansz <58411513+zhangfanfansz@users.noreply.github.com>